Create RockPaperScissors
This was a tick tac toe game I programmed back in early October of 2017. I had to learn how to use methods and arrays and learn about return types and that sort of thing in order to do the project since my class hadn't gotten to any of that yet.