Create QwertyCode
This program is intended to take an input from the user and pass it through a code key where each alphabet letter corresponds to the order of letters on the keyboard. This is a one-way conversion program, and I plan on branching this file to create a program that can reverse it as well.